Moved the Movie model to its own file
