+++ movie-showcase/src/models/movie.ts
@@ -1,0 +1,6 @@
+export type Movie = {
+    Title: string;
+    Year: number;
+    Plot: string;
+    Poster: string;
+}
